Refactor ast to split parse and eval
