+++ src/rust/expression/ast/eval.rs
@@ -1,291 +1,18 @@
+use std::cmp::min;
 use std::collections::HashMap;
-use std::fmt;
 
 use pyo3::exceptions::{
     PyAttributeError, PyIndexError, PyKeyError, PyTypeError, PyZeroDivisionError,
 };
-use pyo3::types::{PyBool, PyDict, PyInt, PyList, PyString, PyTuple};
-use pyo3::{prelude::*, BoundObject, IntoPyObjectExt};
+use pyo3::prelude::*;
+use pyo3::types::{PyDict, PyTuple};
 
 use crate::catalog::XCatalog;
-use crate::markup::tokens::{ToHtml, XNode};
+use crate::expression::ast::parse::parse;
+use crate::expression::{parser::tokenize, tokens::Operator};
+use crate::markup::tokens::ToHtml;
 
-use super::tokens::Operator;
-
-trait Truthy {
-    fn is_truthy(&self) -> bool;
-}
-
-#[pyclass]
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum LiteralKey {
-    Int(isize),
-    Str(String),
-    Uuid(String),
-}
-
-impl fmt::Display for LiteralKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            LiteralKey::Int(v) => write!(f, "{}", v),
-            LiteralKey::Str(v) => write!(f, "\"{}\"", v),
-            LiteralKey::Uuid(v) => write!(f, "\"{}\"", v),
-        }
-    }
-}
-
-impl LiteralKey {
-    fn downcast<'py>(value: Bound<'py, PyAny>) -> Result<Self, PyErr> {
-        if let Ok(v) = value.downcast::<PyInt>() {
-            return Ok(LiteralKey::Int(v.extract::<isize>()?));
-        } else if let Ok(v) = value.downcast::<PyString>() {
-            return Ok(LiteralKey::Str(v.to_string()));
-        } else if value.downcast::<PyAny>()?.get_type().name()? == "UUID" {
-            let uuid_str = value.getattr("hex")?;
-            Ok(LiteralKey::Uuid(uuid_str.to_string()))
-        } else {
-            let err: PyErr = PyTypeError::new_err(format!("Can't use {:?} as LiteralKey", value));
-            return Err(err);
-        }
-    }
-    fn into_py<'py>(&self, py: Python<'py>) -> pyo3::Bound<'py, PyAny> {
-        match self {
-            LiteralKey::Int(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            LiteralKey::Uuid(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            LiteralKey::Str(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-        }
-    }
-}
-
-impl TryFrom<Literal> for LiteralKey {
-    type Error = PyErr;
-
-    fn try_from(lit: Literal) -> Result<Self, Self::Error> {
-        match lit {
-            Literal::Int(i) => Ok(LiteralKey::Int(i)),
-            Literal::Str(s) => Ok(LiteralKey::Str(s)),
-            Literal::Uuid(u) => Ok(LiteralKey::Uuid(u)),
-            _ => Err(PyTypeError::new_err(format!(
-                "Unsupported literal type for key {:?}",
-                lit
-            ))),
-        }
-    }
-}
-
-#[derive(Debug, IntoPyObject)]
-pub struct PyObj {
-    obj: Py<PyAny>,
-}
-
-impl PyObj {
-    pub fn new(obj: Py<PyAny>) -> Self {
-        PyObj { obj }
-    }
-    pub fn obj(&self) -> &Py<PyAny> {
-        &self.obj
-    }
-}
-
-impl Clone for PyObj {
-    fn clone(&self) -> Self {
-        Python::with_gil(|py| PyObj {
-            obj: self.obj.clone_ref(py),
-        })
-    }
-}
-
-#[derive(Debug, Clone, IntoPyObject)]
-pub enum Literal {
-    Bool(bool),
-    Int(isize),
-    Str(String),
-    Uuid(String), // Uuid type does not support IntoPyObject
-    XNode(XNode),
-    List(Vec<Literal>),
-    Dict(HashMap<LiteralKey, Literal>),
-    Callable(String), // the name of the callable
-    Object(PyObj),
-}
-
-impl Literal {
-    fn downcast<'py>(value: Bound<'py, PyAny>) -> Result<Self, PyErr> {
-        if let Ok(v) = value.downcast::<PyString>() {
-            return Ok(Literal::Str(v.to_string()));
-        } else if let Ok(v) = value.downcast::<PyBool>() {
-            return Ok(Literal::Bool(v.extract::<bool>()?));
-        } else if let Ok(v) = value.downcast::<PyInt>() {
-            return Ok(Literal::Int(v.extract::<isize>()?));
-        } else if let Ok(v) = value.extract::<XNode>() {
-            return Ok(Literal::XNode(v));
-        } else if let Ok(seq) = value.downcast::<PyList>() {
-            let mut items = Vec::with_capacity(seq.len());
-            for item in seq.iter() {
-                items.push(Literal::downcast(item)?);
-            }
-            Ok(Literal::List(items))
-        } else if let Ok(dict) = value.downcast::<PyDict>() {
-            let mut map = HashMap::new();
-            for (k, v) in dict {
-                let key: LiteralKey = LiteralKey::downcast(k)?;
-                let val: Literal = Literal::downcast(v)?;
-                map.insert(key, val);
-            }
-            Ok(Literal::Dict(map))
-        } else if value.downcast::<PyAny>()?.get_type().name()? == "UUID" {
-            let uuid_str = value.getattr("hex")?;
-            Ok(Literal::Uuid(uuid_str.to_string()))
-        } else {
-            let o: Py<PyAny> = value.extract()?;
-            Ok(Literal::Object(PyObj::new(o)))
-        }
-    }
-    fn into_py<'py>(&self, py: Python<'py>) -> pyo3::Bound<'py, PyAny> {
-        match self {
-            Literal::Bool(v) => v // wtf
-                .into_pyobject(py)
-                .unwrap()
-                .unbind()
-                .into_bound_py_any(py)
-                .unwrap(),
-            Literal::Uuid(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            Literal::Int(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            Literal::Str(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            Literal::XNode(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            Literal::List(v) => v.clone().into_pyobject(py).unwrap().into_any(),
-            Literal::Callable(v) => v.clone().into_pyobject(py).unwrap().into_any(), // wrong!
-            Literal::Object(v) => v.clone().obj.into_pyobject(py).unwrap().into_any(),
-            Literal::Dict(map) => {
-                let dict = PyDict::new(py);
-                for (k, v) in map {
-                    dict.set_item(
-                        k.clone().into_pyobject(py).unwrap().into_any(),
-                        v.into_py(py),
-                    )
-                    .unwrap();
-                }
-                dict.into_any()
-            }
-        }
-    }
-}
-
-impl Truthy for Literal {
-    fn is_truthy(&self) -> bool {
-        match self {
-            Literal::Bool(bool) => bool.clone(),
-            Literal::Int(i) => *i != 0,
-            Literal::Str(s) => !s.is_empty(),
-            Literal::Uuid(_) => true,
-            Literal::XNode(_) => true,
-            Literal::Callable(_) => true,
-            Literal::Object(o) => {
-                Python::with_gil(|py| {
-                    match o
-                        .obj()
-                        .into_pyobject(py)
-                        .unwrap()
-                        .call_method("__bool__", (), None)
-                    {
-                        Ok(b) => b.extract::<bool>().unwrap(),
-                        Err(_) => false, // or panic/log
-                    }
-                })
-            }
-            Literal::List(items) => !items.is_empty(),
-            Literal::Dict(d) => !d.is_empty(),
-        }
-    }
-}
-
-impl ToHtml for Literal {
-    fn to_html<'py>(
-        &self,
-        py: Python<'py>,
-        catalog: &XCatalog,
-        params: Bound<'py, PyDict>,
-        globals: Bound<'py, PyDict>,
-    ) -> PyResult<String> {
-        debug!("Rendering {:?}", self);
-        match self {
-            Literal::Bool(b) => Ok(format!("{}", b)),
-            Literal::Int(i) => Ok(format!("{}", i)),
-            Literal::Str(s) => Ok(format!("{}", s)),
-            Literal::Callable(s) => Ok(format!("{}()", s)),
-            Literal::Uuid(uuid) => Ok(format!("{}", uuid)),
-            Literal::List(l) => {
-                let mut out = String::new();
-                for item in l {
-                    out.push_str(
-                        item.to_html(py, catalog, params.clone(), globals.clone())?
-                            .as_str(),
-                    );
-                }
-                Ok(out)
-            }
-            Literal::Dict(d) => {
-                let mut out = String::new();
-                out.push_str("<dl>");
-                for (k, item) in d {
-                    out.push_str("<dt>");
-                    out.push_str(format!("{}", k).as_str());
-                    out.push_str("</dt>");
-                    out.push_str("<dt>");
-                    out.push_str(
-                        item.to_html(py, catalog, params.clone(), globals.clone())?
-                            .as_str(),
-                    );
-                    out.push_str("</dt>");
-                }
-                out.push_str("</dl>");
-                Ok(out)
-            }
-            Literal::Object(o) => Ok(format!(
-                "{}",
-                Python::with_gil(|py| {
-                    match o
-                        .obj()
-                        .into_pyobject(py)
-                        .unwrap()
-                        .call_method("__repr__", (), None)
-                    {
-                        Ok(b) => b.extract::<String>().unwrap(),
-                        Err(_) => "<PyObject>".to_string(),
-                    }
-                })
-            )),
-            Literal::XNode(n) => catalog.render_node(py, &n, params.clone(), globals.clone()),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum AST {
-    Variable(String),
-    Literal(Literal),
-    Binary {
-        left: Box<AST>,
-        op: Operator,
-        right: Box<AST>,
-    },
-    FieldAccess(Box<AST>, String),
-    IndexAccess(Box<AST>, Box<AST>),
-    CallAccess {
-        left: Box<AST>,
-        args: Vec<AST>,
-        kwargs: HashMap<String, AST>,
-    },
-    IfStatement {
-        condition: Box<AST>,
-        then_branch: Box<AST>,
-        else_branch: Option<Box<AST>>,
-    },
-    ForStatement {
-        ident: String,
-        iterable: Box<AST>,
-        body: Box<AST>,
-    },
-}
+use super::model::{Literal, LiteralKey, Truthy, AST};
 
 fn eval_add(l: Literal, r: Literal) -> PyResult<Literal> {
     match (l, r) {
@@ -699,4 +426,22 @@ fn wrap_params<'py>(
         result.set_item(key.into_py(py), value.into_py(py))?;
     }
     Ok(result)
+}
+
+pub fn eval_expression<'py>(
+    py: Python<'py>,
+    expression: &str,
+    catalog: &XCatalog,
+    params: Bound<'py, PyDict>,
+    globals: Bound<'py, PyDict>,
+) -> Result<Literal, PyErr> {
+    info!(
+        "Evaluating expression {}...",
+        &expression[..min(expression.len(), 24)]
+    );
+    let params_ast = cast_params(params)?;
+    let token = tokenize(expression)?;
+    let ast = parse(&[token])?;
+    let global_params = cast_params(globals)?;
+    eval_ast(py, &ast, catalog, &params_ast, &global_params)
 }
